Fix ACTION_DEFAULT_SUBSCRIPTION_CHANGED is not sent after power on
Some components depend on ACTION_DEFAULT_SUBSCRIPTION_CHANGED to
initialize the default SUB after UE boot up, such as TelephonyRegistry
When subId is equal with defaultSubId, for adding the first SUB
record info, the subId is active SUB so that the default fallback SUB
is not set, for adding the next SUB record, defaultSubId is valid for
ever, also will not set.

To fix it, we should froce invoking setDefaultFallbackSubId once subId
is equal with defaultSubId.

Change-Id: Ibd3aad0629f989606483b326da3bbf058c77501c
CRs-Fixed: 813620